Bootstrap the IPR database if empty
If the IPR database is new, then it needs to be bootstrapped with users and data.  This is accomplished by putting the .dump file into the persistent storage and loading that.

Additionally reorganized the kubernetes .yaml files into directories indicating their usages.  This makes it harder to load all the files with kubectl, but the organization is helping with development.